requirements.txt
annotated-types==0.7.0
azure-core==1.30.2
azure-identity==1.17.1
azure-keyvault-secrets==4.8.0
babel==2.17.0
black==25.1.0
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.8
ghp-import==2.1.0
graphviz==0.20.3
griffe==0.47.0
idna==3.7
iniconfig==2.0.0
isodate==0.6.1
Jinja2==3.1.4
joblib==1.4.2
Markdown==3.6
MarkupSafe==2.1.5
mergedeep==1.3.4
mkdocs==1.6.0
mkdocs-autorefs==1.0.1
mkdocs-get-deps==0.2.0
mkdocs-material==9.6.3
mkdocs-material-extensions==1.3.1
mkdocstrings==0.24.3
mkdocstrings-python==1.10.0
msal==1.29.0
msal-extensions==1.2.0
mypy-extensions==1.0.0
packaging==24.1
paginate==0.5.7
pathspec==0.12.1
platformdirs==4.2.2
pluggy==1.5.0
portalocker==2.10.0
psycopg2-binary==2.9.9
pycparser==2.22
pydantic==2.7.4
pydantic_core==2.18.4
Pygments==2.19.1
PyJWT==2.8.0
pymdown-extensions==10.8.1
pytest==8.3.4
python-dateutil==2.9.0.post0
python-gnupg==0.5.2
PyYAML==6.0.1
pyyaml_env_tag==0.1
regex==2024.11.6
requests==2.32.3
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
six==1.16.0
typing_extensions==4.12.2
urllib3==2.2.2
watchdog==4.0.1
